compress: Handle case of pause() during compress_write()
If stream is paused() during a compress_write() the poll()
will return 0 or -EBADFD, or the write() will return -EBADFD.
This is not an error so compress_write() should not pass an
error code to the caller. It should abort the write and then
return the number of bytes it had written up to that point.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>