compress: Allow client to pass "don't care" for fragment size/number
Client cannot know what is a sensible fragment size and count
for the device unless it has prior knowledge of the precise
hardware. To allow for generic clients not hardcoded for particular
hardware, this change allows fragment_size and fragments to be
passed as 0, meaning "don't care" and tinycompress will choose
default values read from the driver.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>